+++ code/debug.h
@@ -38,7 +38,7 @@ void clrLED(void);
 void togLED(void);
 void USART_SendChar(char c);
 void USART_SendString(char* StringPtr);
-void USART_SendInt(int num);
+void USART_SendInt(uint16_t num);
 void USART_NewLine(void);
 
 #endif //File guard
